+++ src/implementations/react/src/hig-react.js
@@ -1,28 +1,28 @@
 export { default as Avatar } from './adapters/AvatarAdapter';
 export { default as Button } from './adapters/ButtonAdapter';
-export { colors } from 'hig-vanilla';
 export { default as Checkbox } from './elements/components/FormElements/Checkbox';
+export { colors } from 'hig-vanilla';
 export { default as Dropdown } from './elements/components/FormElements/Dropdown';
-export { default as SideNav } from './elements/components/GlobalNav/SideNav';
+export { default as Flyout } from './elements/components/Flyout';
 export { default as GlobalNav } from './elements/components/GlobalNav/GlobalNav';
+export { default as Grid } from './adapters/GridAdapter';
+export { default as GridItem } from './elements/components/GridItem';
 export { default as Icon } from './adapters/IconAdapter';
 export { default as IconButton } from './adapters/IconButtonAdapter';
-export { default as Flyout } from './elements/components/Flyout';
 export { default as Modal } from './elements/components/Modal';
 export { default as Option } from './adapters/FormElements/OptionAdapter';
 export { default as PasswordField } from './elements/components/FormElements/PasswordField';
 export { default as RadioButton } from './adapters/FormElements/RadioButtonAdapter';
 export { default as Range } from './elements/components/FormElements/Range';
 export { default as RichText } from './elements/components/RichText';
-export { default as Spacer } from './adapters/SpacerAdapter';
+export { default as SideNav } from './elements/components/GlobalNav/SideNav';
 export { sizes } from 'hig-vanilla';
-export { default as TextArea } from './elements/components/FormElements/TextArea';
-export { default as TextField } from './elements/components/FormElements/TextField';
-export { default as TextLink } from './adapters/TextLinkAdapter';
-export * from './elements/components/Typography';
-export { default as Table } from  './adapters/TableAdapter';
 export { default as Slot } from './adapters/SlotAdapter';
 export { default as SlotHeadCell } from './elements/components/SlotHeadCell';
 export { default as SlotCell } from './elements/components/SlotCell';
-export { default as Grid } from './adapters/FormElements/GridAdapter';
-export { default as GridItem } from './elements/components/FormElements/GridItem';
+export { default as Spacer } from './adapters/SpacerAdapter';
+export { default as TextArea } from './elements/components/FormElements/TextArea';
+export { default as TextField } from './elements/components/FormElements/TextField';
+export { default as TextLink } from './adapters/TextLinkAdapter';
+export { default as Table } from  './adapters/TableAdapter';
+export * from './elements/components/Typography';
